Cleanup Game test move to Controller tests
Game test contained a lot of tests that where actually testing the
AlienController for example.  These needed to be moved to their respective
ControllerTests.

<interface, organisation, testing
